backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
anthropic==0.18.1
python-multipart==0.0.6
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
markdown==3.5.2
python-frontmatter==1.1.0
python-docx==1.1.0
httpx==0.26.0
pytest==7.4.4
pytest-asyncio==0.23.3

# Audio generation
elevenlabs==1.0.0
pydub==0.25.1

# Image processing
Pillow==10.2.0
